+++ config/config.yaml.example
@@ -1,0 +1,22 @@
+server:
+  port: "8080"
+  read_timeout: 30s
+  write_timeout: 30s
+
+database:
+  host: localhost
+  port: 3306
+  user: user
+  password: pass
+  dbname: smartmail
+
+gmail:
+  client_id: your-client-id
+  client_secret: your-client-secret
+  refresh_token: your-refresh-token
+  user_email: you@example.com
+  use_imap: false
+
+scheduler:
+  interval_minutes: 5
+  max_retries: 3
