+++ internal/router/router.go
@@ -1,0 +1,36 @@
+package router
+
+import (
+	"fmt"
+	"time"
+
+	"github.com/gin-gonic/gin"
+
+	"smart-mail-relay-go/internal/handler"
+)
+
+// SetupRouter configures the Gin router with routes and middleware
+func SetupRouter(h *handler.Handlers) *gin.Engine {
+	gin.SetMode(gin.ReleaseMode)
+	r := gin.New()
+	r.Use(gin.Recovery())
+	r.Use(loggerMiddleware())
+	h.SetupRoutes(r)
+	return r
+}
+
+func loggerMiddleware() gin.HandlerFunc {
+	return gin.LoggerWithFormatter(func(param gin.LogFormatterParams) string {
+		return fmt.Sprintf("%s - [%s] \"%s %s %s %d %s \"%s\" %s\"\n",
+			param.ClientIP,
+			param.TimeStamp.Format(time.RFC1123),
+			param.Method,
+			param.Path,
+			param.Request.Proto,
+			param.StatusCode,
+			param.Latency,
+			param.Request.UserAgent(),
+			param.ErrorMessage,
+		)
+	})
+}
